Remove provider
