+++ storage/postgres/storage.go
@@ -17,9 +17,45 @@
 package postgres
 
 import (
+	"os"
+	"os/signal"
+	"syscall"
+
+	"github.com/Peripli/service-manager/env"
 	"github.com/Peripli/service-manager/storage"
+	"github.com/Sirupsen/logrus"
 	"github.com/jmoiron/sqlx"
+	_ "github.com/lib/pq"
 )
+
+func init() {
+	closeChan := make(chan os.Signal)
+	signal.Notify(closeChan, os.Interrupt, syscall.SIGTERM)
+
+	uri, ok := env.Get("storage.uri").(string)
+	if !ok {
+		logrus.Panicf("Could not open connection for provided uri %s from postgres storage provider", uri)
+	}
+	db, err := sqlx.Open("postgres", uri)
+	if err != nil {
+		logrus.Panicf("Could not connect to PostgreSQL storage" + err.Error())
+	}
+
+	go func() {
+		<-closeChan
+		logrus.Debug("Received close for postgres storage")
+		close(closeChan)
+		if err := db.Close(); err != nil {
+			logrus.Panic(err)
+		}
+	}()
+	storageToRegister, err := newStorage(db)
+	if err != nil {
+		logrus.Panicf("Cannot initialize postgres storage. Error : %v", err)
+	}
+	logrus.Debug("Initialized PostgreSQL storage")
+	storage.Register("postgres", storageToRegister)
+}
 
 type postgresStorage struct {
 }
